+++ src/types/index.ts
@@ -1,0 +1,3 @@
+export * from './expressionParts';
+export * from './evaluator';
+export * from './paths';
