+++ src/types/paths.ts
